Re-organize project solution
