+++ Software/CT.BusinessLogic/Interfaces/ICANanalys.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace CT.BusinessLogic.Interfaces
+{
+    public interface ICANanalys
+    {
+
+    }
+}
