+++ Software/CT.BusinessLogic/Interfaces/ISerialService.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CT.BusinessLogic.Interfaces
+{
+    public interface ISerialService
+    {
+        Task<string> GetSerialValue();
+    }
+}
