+++ Software/CT.BusinessLogic/Services/CANanalys.cs
@@ -2,10 +2,11 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using CT.BusinessLogic.Interfaces;
 
-namespace car_telemetry.Data
+namespace CT.BusinessLogic.Services
 {
-    public class CANanalys
+    public class CANanalys : ICANanalys
     {
         //дані будуть йти в форматі CAN пакетів через серіал: 
         //"0x0F6 8 0x8E 0x87 0x32 0xFA 0x26 0x8E 0xBE 0x86"
