+++ Software/CT.Web/car-telemetry.csproj
@@ -9,6 +9,10 @@
     <PackageReference Include="System.IO.Ports" Version="5.0.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\CT.BusinessLogic\CT.BusinessLogic.csproj" />
+  </ItemGroup>
+
   <ProjectExtensions><VisualStudio><UserProperties appsettings_1json__JsonSchema="" /></VisualStudio></ProjectExtensions>
 
 </Project>
