+++ Software/CT.Web/wwwroot/css/open-iconic/FONT-LICENSE
