+++ Software/CT.Web/wwwroot/css/open-iconic/ICON-LICENSE
