+++ Software/CT.Web/wwwroot/css/open-iconic/README.md
