+++ Software/CT.Web/wwwroot/css/open-iconic/font/fonts/open-iconic.otf
