+++ Software/CT.Web/wwwroot/images/passangerOpenedDoor.png
